KEA-147 added constructor to test
